Revert "[nrf fromlist] wifi: Fix fractional part of Tx rate by converting to float"
This reverts commit ae694c3e15d6fa771c6303615dab81ae4b4a3a5a.

+++ include/zephyr/net/wifi_mgmt.h
@@ -672,7 +672,7 @@ struct wifi_iface_status {
 	/** is TWT capable? */
 	bool twt_capable;
 	/** The current 802.11 PHY TX data rate (in Mbps) */
-	float current_phy_tx_rate;
+	int current_phy_tx_rate;
 };
 
 /** @brief Wi-Fi power save parameters */
